requirements.txt
-e git+https://git@github.com/AI4SCR/ai4bmr-datasets.git@main#egg=ai4bmr_datasets
annotated-types==0.7.0
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cramjam==2.11.0
cycler==0.12.1
et_xmlfile==2.0.0
fastparquet==2025.12.0
fonttools==4.61.1
fsspec==2026.1.0
geopandas==1.1.2
idna==3.11
imagecodecs==2026.1.14
ImageIO==2.37.2
kiwisolver==1.4.9
lazy_loader==0.4
loguru==0.7.3
matplotlib==3.10.8
networkx==3.6.1
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pillow==12.1.0
pyarrow==21.0.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
scikit-image==0.26.0
scipy==1.17.0
setuptools==80.10.1
shapely==2.1.2
six==1.17.0
tifffile==2026.1.28
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wheel==0.46.3
